n x m clockwise matrix printing
This program was a daily coding challenge.  Given an N X M matrix, print the values of the matrix in a clockwise spiral pattern.  This program initializes an N X M matrix based on user input, and then prints the values in order of a clockwise spiral traversal of the matrix.